JAV-256 persisted events to database

+++ saga-core/src/main/java/io/servicecomb/saga/core/SagaLog.java
@@ -20,5 +20,5 @@ public interface SagaLog {
 
   void offer(SagaEvent sagaEvent);
 
-  int size();
+  long size();
 }
